Allow for missing return value
Allowing one of the return values to be missing from the speedtest return data. Apparently if Ookla's results servers fail, the speedtest will still run but there will be no URL returned. This was seen when their results servers started returning a 500 error. Since we don't care about the results URL, and the speedtest data is still good, we'll allow it.